+++ SamplesApp/SamplesApp/Features/VideoChat/VideoChatConfigViewController.swift
@@ -146,9 +146,7 @@ class VideoChatConfigViewController: UITableViewController {
     // 入力された設定を元にSoraへ接続を行います。
     // ビデオチャットアプリでは複数のユーザーが同時に配信を行う必要があるため、
     // role 引数には .sendrecv を指定します。
-    VideoChatSoraSDKManager.shared.connect(
-      with: configuration
-    ) { [weak self] error in
+    SoraSDKManager.shared.connect(configuration: configuration) { [weak self] error in
       // 接続処理が終了したので false にします。
       self?.isConnecting = false
 
@@ -157,7 +155,7 @@ class VideoChatConfigViewController: UITableViewController {
         // この場合は、エラー表示をユーザーに返すのが親切です。
         // なお、このコールバックはメインスレッド以外のスレッドから呼び出される可能性があるので、
         // UI操作を行う際には必ずDispatchQueue.main.asyncを使用してメインスレッドでUI処理を呼び出すようにしてください。
-        logger.warning("[sample] VideoChatSoraSDKManager connection error: \(error)")
+        logger.warning("[sample] SoraSDKManager connection error: \(error)")
         DispatchQueue.main.async {
           let alertController = UIAlertController(
             title: "接続に失敗しました",
@@ -169,7 +167,7 @@ class VideoChatConfigViewController: UITableViewController {
         }
       } else {
         // errorがnilの場合は、接続に成功しています。
-        logger.info("[sample] VideoChatSoraSDKManager connected.")
+        logger.info("[sample] SoraSDKManager connected.")
 
         // 次の配信画面に遷移します。
         // なお、このコールバックはメインスレッド以外のスレッドから呼び出される可能性があるので、
@@ -178,7 +176,7 @@ class VideoChatConfigViewController: UITableViewController {
           guard let self else { return }
 
           if !shouldEnableCameraOnConnect,
-            let mediaChannel = VideoChatSoraSDKManager.shared.currentMediaChannel
+            let mediaChannel = SoraSDKManager.shared.currentMediaChannel
           {
             // 映像ハードミュートを有効にします
             // setVideoHardMute は async メソッドのため Task 内で実行します